+++ kudos/Domain/Protocols/AddAccomplishmentUseCaseProtocol.swift
@@ -1,0 +1,3 @@
+protocol AddAccomplishmentUseCaseProtocol {
+    func execute(text: String, color: String) throws
+}
