+++ kudos/Domain/Protocols/AddPhotoAccomplishmentUseCaseProtocol.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol AddPhotoAccomplishmentUseCaseProtocol {
+    func execute(photoData: Data, caption: String?, color: String) throws
+}
