+++ kudos/Domain/Protocols/DeleteAccomplishmentUseCaseProtocol.swift
@@ -1,0 +1,3 @@
+protocol DeleteAccomplishmentUseCaseProtocol {
+    func execute(_ accomplishment: AccomplishmentItem) throws
+}
