+++ kudos/Domain/Protocols/GetAccomplishmentsUseCaseProtocol.swift
@@ -1,0 +1,3 @@
+protocol GetAccomplishmentsUseCaseProtocol {
+    func execute() throws -> [AccomplishmentItem]
+}
