+++ kudos/Domain/useCases/AddPhotoAccomplishmentUseCase.swift
@@ -1,9 +1,5 @@
 import Foundation
 
-protocol AddPhotoAccomplishmentUseCaseProtocol {
-    func execute(photoData: Data, caption: String?, color: String) throws
-}
-
 final class AddPhotoAccomplishmentUseCase: AddPhotoAccomplishmentUseCaseProtocol {
     private let repository: AccomplishmentRepositoryProtocol
